requirements.txt
yt-dlp
spotipy
python-dotenv